# docs/user-api-docs-after.yaml
openapi: 3.0.3
info:
  title: CEC USER API Documentation
  version: 1.0.12
...
paths:
...
  /api/inquiry/{id}:
    get:
      tags:
        - 3-2. 유저 프로필 / 1:1 문의하기
      summary: 문의글 상세 조회
      description: 본인이 작성한 문의글 + 답변 내용
      operationId: getInquiry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInquiryResponse'
    put:
      tags:
        - 3-2. 유저 프로필 / 1:1 문의하기
      summary: 문의글 수정
      description: 본인이 작성한 문의글을 수정합니다.
      operationId: updateInquiry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InquiryRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseLong'
    delete:
      tags:
        - 3-2. 유저 프로필 / 1:1 문의하기
      summary: 문의글 삭제
      description: 본인이 작성한 문의글을 삭제합니다.
      operationId: deleteInquiry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
  /api/equipments/cart:
    get:
      tags:
        - 1-3. 대여 신청 / 장비
      summary: 장비 장바구니 조회
      operationId: getCartItems
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListEquipmentResponse'
    post:
      tags:
        - 1-3. 대여 신청 / 장비
      summary: 장비 장바구니 추가
      operationId: addToCart
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
                format: int64
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
...
  /api/inquiry:
    get:
      tags:
        - 3-2. 유저 프로필 / 1:1 문의하기
      summary: 내 문의글 목록 조회
      description: 로그인한 사용자의 모든 문의글을 페이지네이션과 함께 조회합니다.
      operationId: getMyInquiries
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            default: createdAt
        - name: sortDirection
          in: query
          required: false
          schema:
            type: string
            default: DESC
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInquiryListResponse'
    post:
      tags:
        - 3-2. 유저 프로필 / 1:1 문의하기
      summary: 문의글 작성
      description: 인증된 사용자가 제목, 내용, 유형 등을 입력하여 문의글을 작성합니다.
      operationId: createInquiry
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InquiryRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseLong'
  /api/equipments/action:
    patch:
      tags:
        - 1-3. 대여 신청 / 장비
      summary: 장비 상태 변경 요청 (대여/반납 요청 및 취소)
      description: 날짜는 대여 요청 시에만 필요
      operationId: handleEquipmentAction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentActionRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
...
  /api/equipments:
    get:
      tags:
        - 1-3. 대여 신청 / 장비
      summary: 장비 목록 조회
      operationId: getEquipments
      parameters:
        - name: categoryId
          in: query
          description: 카테고리 ID로 정렬
          required: false
          schema:
            type: string
            description: 카테고리 ID로 정렬
            example: 1
          example: 1
        - name: searchKeyword
          in: query
          description: 검색 키워드 (모델명, 대여자명, 일련번호, 카테고리명)
          required: false
          schema:
            type: string
            description: 검색 키워드 (모델명, 대여자명, 일련번호, 카테고리명)
            example: SONY
          example: SONY
        - name: searchType
          in: query
          description: 검색 타입 (ALL, MODEL_NAME, CATEGORY_NAME, SERIAL_NUMBER, RENTER_NAME)
          required: false
          schema:
            type: string
            enum:
              - ALL
              - MODEL_NAME
              - CATEGORY_NAME
              - RENTER_NAME
          example: ALL
        - name: status
          in: query
          description: >-
            장비 상태 (ALL, AVAILABLE, IN_USE, BROKEN, RENTAL_PENDING,
            RETURN_PENDING 중 선택)
          required: false
          schema:
            type: string
            enum:
              - ALL
              - AVAILABLE
              - IN_USE
              - BROKEN
              - RENTAL_PENDING
              - RETURN_PENDING
          example: AVAILABLE
        - name: sortBy
          in: query
          description: 정렬할 필드 (id, createdAt, name 등)
          required: false
          schema:
            type: string
            enum:
              - NAME
              - ID
          example: id
        - name: page
          in: query
          description: 페이지 번호 (기본값 0)
          required: false
          schema:
            type: integer
            description: 페이지 번호 (기본값 0)
            example: 0
          example: 0
        - name: size
          in: query
          description: 페이지당 크기 (기본값 = 10)
          required: false
          schema:
            type: integer
            description: 페이지당 크기 (기본값 = 10)
            example: 10
          example: 10
        - name: sortDirection
          in: query
          description: 정렬 방법
          required: false
          schema:
            type: string
            enum:
              - ASC
              - DESC
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseEquipmentListResponse'
  /api/equipments/{id}:
    get:
      tags:
        - 1-3. 대여 신청 / 장비
      summary: 장비 단일 조회
      operationId: getEquipment
      parameters:
        - name: id
          in: path
          description: 장비 ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseEquipmentResponse'
...
  /api/equipment-categories/{id}:
    get:
      tags:
        - 1-1. 대여 신청 / 장비 분류
      summary: 카테고리 단일 조회
      operationId: getCategoryById
      parameters:
        - name: id
          in: path
          description: 조회할 장비 카테고리의 ID
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseEquipmentCategoryResponse'
...
  /api/classroom/{id}/schedule:
    get:
      tags:
        - 1-4. 대여 신청 / 강의실
      summary: 휴일 + 특강 + 수업 + 대여 시간표 조회
      description: |-
        **일정이 시작되는 시간순으로 정렬해서 조회됩니다.**
      operationId: getSchedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListScheduleResponse'
components:
  schemas:
...
    AuthorResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nickname:
          type: string
        role:
          type: string
          enum:
            - ROLE_USER
            - ROLE_ADMIN
            - ROLE_SUPER_ADMIN
        imageUrl:
          type: string
...
    InquiryRequest:
      required:
        - content
        - title
        - type
      type: object
      properties:
        title:
          type: string
          description: 문의 제목
          example: 장비 대여 관련 문의
        content:
          type: string
          description: 문의 내용
          example: 장비 대여 예약이 되지 않습니다.
        attachments:
          type: array
          items:
            type: string
            description: 첨부파일 URL
            example: https://example.com/image.png, https://example.com/docs.md
        type:
          type: string
          description: 문의 유형
          example: RENTAL
          enum:
            - RENTAL
            - RETURN
            - SYSTEM_ERROR
            - EQUIPMENT_BROKEN
            - PENALTY
            - ETC
    ApiResponseVoid:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        data:
          type: object
...
    ApiResponseLong:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        data:
          type: integer
          format: int64
...
    EquipmentActionRequest:
      type: object
      properties:
        action:
          type: string
          description: |-
            RENT_REQUEST: 대여 요청
            RENT_CANCEL: 대여 요청 취소
            RETURN_REQUEST: 반납 요청
            RETURN_CANCEL: 반납 요청 취소
          enum:
            - RENT_REQUEST
            - RENT_CANCEL
            - RETURN_REQUEST
            - RETURN_CANCEL
        ids:
          type: array
          description: 장비 ID 목록
          example:
            - 1
            - 2
            - 3
          items:
            type: integer
            description: 장비 ID 목록
            format: int64
        startAt:
          type: string
          description: 요청 시작일 (대여 요청 시에만 필요)
          format: date-time
          nullable: true
        endAt:
          type: string
          description: 요청 종료일 (대여 요청 시에만 필요)
          format: date-time
          nullable: true
...
    ApiResponseEquipmentListResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        data:
          $ref: '#/components/schemas/EquipmentListResponse'
    EquipmentListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/EquipmentResponse'
        pageable:
          $ref: '#/components/schemas/PageableInfo'
    EquipmentResponse:
      type: object
      properties:
        modelName:
          type: string
          description: 장비 모델 이름
          example: SONY
        startRentAt:
          type: string
          description: 대여 시작 시간
          nullable: true
          format: date-time
        endRentAt:
          type: string
          description: 대여 종료 시간
          nullable: true
          format: date-time
        renterName:
          type: string
          description: 대여자 이름 또는 사용하는 수업 이름 (대여하지 않은 경우 null)
          nullable: true
          example: 홍길동
        status:
          type: string
          example: AVAILABLE, IN_USE, BROKEN, RENTAL_PENDING 중 하나
          enum:
            - AVAILABLE(대여 가능)
            - IN_USE(대여중)
            - BROKEN(파손)
            - RENTAL_PENDING(대여 신청 중)
            - RETURN_PENDING(반납 처리중)
        createdAt:
          type: string
          description: 장비 등록일
          format: date-time
        imageUrl:
          type: string
          description: 장비 이미지 URL
          example: https://example.com/images/equipment.jpg
    PageableInfo:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
    ApiResponseEquipmentResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        data:
          $ref: '#/components/schemas/EquipmentResponse'
    ApiResponseListEquipmentResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/EquipmentResponse'
...
    ApiResponseEquipmentCategoryResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        data:
          $ref: '#/components/schemas/EquipmentCategoryResponse'
    EquipmentCategoryResponse:
      type: object
      properties:
        id:
          type: integer
          description: 장비 카테고리 ID
          format: int64
          example: 1
        name:
          type: string
          description: 장비 카테고리 이름
          example: 카메라
        englishCode:
          type: string
          description: 장비 카테고리 영어 코드
          example: CAMERA
        maxRentalCount:
          type: integer
          description: 최대 대여 가능 개수
          format: int32
          example: 2
...
    ApiResponseListScheduleResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleResponse'
    ScheduleResponse:
      type: object
      properties:
        type:
          type: string
          description: HOLIDAY(휴일), SPECIAL_LECTURE(특강), LECTURE(수업), RENTAL(사용자가 대여) 중 하나.
          enum:
            - HOLIDAY
            - SPECIAL_LECTURE
            - LECTURE
            - RENTAL
        day:
          type: integer
          description: 요일을 나타냄. 1, 2, 3, 4, 5 중 하나.
          format: int32
        date:
          type: string
          format: date
        startTime:
          type: string
          description: 일정 시작 시간
          example: '13:00'
        endTime:
          type: string
          description: 일정 종료 시간
          example: '13:00'
        color:
          type: string
        content:
          type: string
          description:
            type = HOLIDAY / SPECIAL_LECTURE / LECTURE 이면 일정 이름을 나타냅니다.<br>
            type = RENTAL 이면, 이 값은 항상 "대여 됨" 입니다.
        id:
          type: integer
          description: 강의실의 id 입니다.(대여 취소할때 사용)
          nullable: true
          format: int64
        isRenter:
          type: boolean
          description: 현재 로그인된 사람이 강의실 대여자라면 true(대여 취소할때 사용)
    ApiResponseInquiryResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        data:
          $ref: '#/components/schemas/InquiryResponse'
    InquiryResponse:
      type: object
      properties:
        id:
          type: integer
          description: 문의 ID
          format: int64
          example: 1
        title:
          type: string
          description: 문의 제목
          example: 장비 대여 관련 문의
        content:
          type: string
          description: 문의 내용
          example: 장비 대여 예약이 되지 않습니다.
        attachments:
          type: array
          items:
            type: string
            example: https://example.com/image.png, https://example.com/docs.md
            description: 첨부파일 URL
        type:
          type: string
          description: 문의 유형
          example: RENTAL
          enum:
            - RENTAL
            - RETURN
            - SYSTEM_ERROR
            - BROKEN
            - PENALTY
            - ETC
        status:
          type: string
          description: 답변 상태
          example: NOT_ANSWERED
          enum:
            - NOT_ANSWERED
            - ANSWERED
        answer:
          $ref: '#/components/schemas/InquiryAnswerResponse'
        createdAt:
          type: string
          description: 문의 작성 시간 (ISO-8601)
          example: '2025-06-02T15:34:21'
    InquiryAnswerResponse:
      type: object
      properties:
        content:
          type: string
        author:
          $ref: '#/components/schemas/AuthorResponse'
    ApiResponseInquiryListResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        data:
          $ref: '#/components/schemas/InquiryListResponse'
    InquiryListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/InquiryResponse'
        pageable:
          $ref: '#/components/schemas/PageableInfo'